Connecting to the API - Creating a hook for the summary
