+++ src/components/Summary/index.tsx
@@ -1,31 +1,14 @@
-import { useContext } from "react";
 import { ArrowCircleDown, ArrowCircleUp, CurrencyDollar } from "phosphor-react";
 
-import { TransactionsContext } from "../../contexts/TransactionsContext";
+import { useSummary } from "../../hooks/useSummary";
 import { priceFormatter } from "../../utils/formatter";
 
 import { SummaryCard, SummaryContainer } from "./styles";
 
 
-export function Summary() {
-  const { transactions } = useContext(TransactionsContext);
 
-  const summary = transactions.reduce(
-    (acc, transaction) => {
-      if(transaction.type === 'income') {
-        acc.income += transaction.price;
-        acc.total += transaction.price;
-      } else {
-        acc.outcome += transaction.price;
-        acc.total -= transaction.price;
-      }
-      
-      return acc;
-    }, {
-      income: 0,
-      outcome: 0, 
-      total: 0
-    })
+export function Summary() {
+  const summary = useSummary();
 
   return (
     <SummaryContainer>
